Moved action selection decisions into their own decisions.js file
