+++ gemp-module/gemp-stccg-client/src/main/web/js/gemp-022/gameUi.js
@@ -4,6 +4,7 @@ import Card from './jCards.js';
 import { createCardDiv, createFullCardDiv, getCardDivFromId } from './jCards.js';
 import { NormalCardGroup, PlayPileCardGroup, NormalGameCardGroup, TableCardGroup } from './jCardGroup.js';
 import { animateActionResult, communicateActionResult, getSpacelineIndexFromLocationId } from './actionResults.js';
+import { processDecision } from './decisions.js';
 import GameAnimations from './gameAnimations.js';
 import ChatBoxUI from './chat.js';
 import { openSizeDialog, showLinkableCardTitle, removeFromArray } from "./common.js";
@@ -1184,7 +1185,7 @@ export default class GameTableUI {
 
             let pendingDecision = gameState.pendingDecision;
             if (pendingDecision != null) {
-                this.animations.processDecision(pendingDecision, true);
+                processDecision(pendingDecision, true, this);
                 this.startAnimatingTitle();
             } else {
                 this.animations.updateGameState(false);
@@ -1252,7 +1253,7 @@ export default class GameTableUI {
                     let userDecision = gameState.pendingDecision;
                     if (this.hasDecision === false && userDecision != null && typeof userDecision != "undefined") {
                         this.hasDecision = true;
-                        this.animations.processDecision(userDecision, animate);
+                        processDecision(userDecision, animate, this);
                         this.startAnimatingTitle();
                     }
                 }
@@ -1926,157 +1927,6 @@ export default class GameTableUI {
         $('.ui-dialog :button').blur();
     }
 
-    // Choosing one action to resolve, for example phase actions
-    cardActionChoiceDecision(decision) {
-        var id = decision.decisionId;
-        var text = decision.text;
-        let noPass = decision.noPass; // boolean
-        let selectableCards = decision.displayedCards;
-
-        var that = this;
-
-        if (selectableCards.length == 0 && this.gameSettings.get("autoPass") && !this.replayMode) {
-            that.decisionFunction(id, "");
-            return;
-        }
-
-        var selectedCardIds = new Array();
-        var allCardIds = new Array();
-
-        this.alertText.html(text);
-        // ****CCG League****: Border around alert box
-        this.alertBox.addClass("alert-box-highlight");
-
-        var processButtons = function () {
-            that.alertButtons.html("");
-            if (!noPass && selectedCardIds.length == 0) {
-                that.alertButtons.append("<button id='Pass'>Pass</button>");
-                $("#Pass").button().click(function () {
-                    finishChoice();
-                });
-            }
-            if (selectedCardIds.length > 0) {
-                that.alertButtons.append("<button id='ClearSelection'>Reset choice</button>");
-                that.alertButtons.append("<button id='Done'>Done</button>");
-                $("#Done").button().click(function () {
-                    finishChoice();
-                });
-                $("#ClearSelection").button().click(function () {
-                    resetChoice();
-                });
-            }
-        };
-
-        var finishChoice = function () {
-            that.alertText.html("");
-            // ****CCG League****: Border around alert box
-            that.alertBox.removeClass("alert-box-highlight");
-            that.alertButtons.html("");
-            that.clearSelection();
-            $(".card").each(
-                function () {
-                    var card = $(this).data("card");
-                    if (card.zone == "EXTRA") {
-                        $(this).remove();
-                    }
-                });
-            that.hand.layoutCards();
-            that.decisionFunction(id, "" + selectedCardIds);
-        };
-
-        var resetChoice = function () {
-            selectedCardIds = new Array();
-            that.clearSelection();
-            allowSelection();
-            processButtons();
-        };
-
-        var allowSelection = function () {
-            var hasVirtual = false;
-
-            for (let i = 0; i < selectableCards.length; i++) {
-                let selectableCard = selectableCards[i];
-                let actionId = selectableCard.actionId;
-                let actionText = selectableCard.actionText;
-                let blueprintId = selectableCard.blueprintId;
-                let cardId = selectableCard.cardId;
-                let actionType = selectableCard.actionType;
-                let cardIdElem;
-
-                if (blueprintId === "inPlay") {
-                    // in play, do not add "extra" to card id value when doing lookup
-                    cardIdElem = getCardDivFromId(cardId);
-                    allCardIds.push(cardId);
-                } else {
-                    // not in play, need to add "extra" to card id value when doing lookup
-                    hasVirtual = true;
-                    cardIdElem = getCardDivFromId(`extra${cardId}`);
-                    allCardIds.push(`extra${cardId}`);
-                    let zone = "EXTRA";
-                    // No new cardId - interesting that it's going to retrieve an extra{cardId} but
-                    //                 create one with a regular cardId. Intentional?
-                    let imageUrl = selectableCard.imageUrl;
-                    let noOwner = "";
-                    let noLocationIndex = "";
-                    let upsideDown = false;
-                    let card = new Card(blueprintId, zone, cardId, noOwner, imageUrl, noLocationIndex, upsideDown);
-
-                    let cardDiv = that.createCardDivWithData(card);
-                    $(cardDiv).css({opacity: "0.8"});
-
-                    $("#main").append(cardDiv);
-                }
-
-                if (cardIdElem.data("action") == null) {
-                    cardIdElem.data("action", new Array());
-                }
-                var actions = cardIdElem.data("action");
-                actions.push({actionId: actionId, actionText: actionText, actionType: actionType});
-            }
-
-            if (hasVirtual) {
-                that.hand.layoutCards();
-            }
-
-            that.selectionFunction = function (cardId, event) {
-                // DEBUG: console.log("cardActionChoiceDecision -> allowSelection -> selectionFunction");
-                var cardIdElem = getCardDivFromId(cardId);
-                var actions = cardIdElem.data("action");
-
-                var selectActionFunction = function (actionId) {
-                    selectedCardIds.push(actionId);
-                    if (that.gameSettings.get("autoAccept")) {
-                        finishChoice();
-                    } else {
-                        that.clearSelection();
-                        getCardDivFromId(cardId).addClass("selectedCard");
-                        processButtons();
-                    }
-                };
-
-                // If the only legal action is a card play, perform action automatically by clicking
-                // Otherwise show a drop-down menu with the action options by clicking
-                if (actions.length == 1 &&
-                        (actions[0].actionType == "PLAY_CARD" || actions[0].actionType == "SEED_CARD")) {
-                    selectActionFunction(actions[0].actionId);
-                } else {
-                        // TODO - Bind to right-click?
-                    that.createActionChoiceContextMenu(actions, event, selectActionFunction);
-                }
-            };
-
-            that.attachSelectionFunctions(allCardIds, false);
-        };
-
-        allowSelection();
-        if (!this.replayMode) {
-            processButtons();
-            this.PlayAwaitActionSound();
-        }
-
-        $(':button').blur();
-    }
-
     PlayAwaitActionSound() {
         let audio = new Audio(awaitingActionAudio);
         if(!document.hasFocus() || document.hidden || document.msHidden || document.webkitHidden)
@@ -2131,99 +1981,6 @@ export default class GameTableUI {
         setTimeout(function () { // Delay for Mozilla
             $(document).click(getRidOfContextMenu);
         }, 0);
-    }
-
-    // Choosing one action to resolve, for example required triggered actions
-    actionChoiceDecision(decision) {
-        var id = decision.decisionId;
-        var text = decision.text;
-
-        let displayedCards = decision.displayedCards;
-
-        var that = this;
-
-        var selectedActionIds = new Array();
-
-        this.cardActionDialog
-            .html("<div id='arbitraryChoice'></div>")
-            .dialog("option", "title", text);
-
-        var cardIds = new Array();
-
-        for (let i = 0; i < displayedCards.length; i++) {
-            let displayedCard = displayedCards[i];
-            let blueprintId = displayedCard.blueprintId;
-            let zone = "SPECIAL";
-            let cardId = `temp${i}`;
-            let noOwner = "";
-            let imageUrl = displayedCard.imageUrl;
-            let noLocationIndex = "";
-            let upsideDown = false;
-
-            cardIds.push("temp" + i);
-            let card = new Card(blueprintId, zone, cardId, noOwner, imageUrl, noLocationIndex, upsideDown);
-
-            let cardDiv = this.createCardDivWithData(card, displayedCard.actionText);
-
-            $("#arbitraryChoice").append(cardDiv);
-        }
-
-        var finishChoice = function () {
-            that.cardActionDialog.dialog("close");
-            $("#arbitraryChoice").html("");
-            that.clearSelection();
-            that.decisionFunction(id, "" + selectedActionIds);
-        };
-
-        var resetChoice = function () {
-            selectedActionIds = new Array();
-            that.clearSelection();
-            allowSelection();
-            processButtons();
-        };
-
-        var processButtons = function () {
-            var buttons = {};
-            if (selectedActionIds.length > 0) {
-                buttons["Clear selection"] = function () {
-                    resetChoice();
-                    processButtons();
-                };
-                buttons["Done"] = function () {
-                    finishChoice();
-                };
-            }
-            that.cardActionDialog.dialog("option", "buttons", buttons);
-        };
-
-        var allowSelection = function () {
-            that.selectionFunction = function (cardId) {
-                // DEBUG: console.log("actionChoiceDecision -> allowSelection -> selectionFunction");
-                var actionId = displayedCards[parseInt(cardId.substring(4))].actionId;
-                selectedActionIds.push(actionId);
-
-                that.clearSelection();
-
-                if (this.gameSettings.get("autoAccept")) {
-                    finishChoice();
-                } else {
-                    processButtons();
-                    getCardDivFromId(cardId).addClass("selectedCard");
-                }
-            };
-
-            that.attachSelectionFunctions(cardIds, true);
-        };
-
-        allowSelection();
-        if (!this.replayMode) {
-            processButtons();
-            this.PlayAwaitActionSound();
-        }
-
-        openSizeDialog(this.cardActionDialog);
-        this.arbitraryDialogResize(false);
-        $('.ui-dialog :button').blur();
     }
 
     // Choosing some number of cards, for example to wound
